v0.6
Session and login implementation

+++ src/deleteTask.php
@@ -1,8 +1,10 @@
 <?php
 
+session_start();
+
 $mysql = new mysqli('127.0.0.1', 'root', '', 'control_tiempo');
 
-$delete = "DELETE FROM task WHERE ID = ".$_POST['id']."";
+$delete = "DELETE FROM task WHERE ID = ".$_POST['id']." AND USER_ID = ".$_SESSION['ID_USER']."";
 
 if ( $mysql -> query ( $delete ) === TRUE ) {
